# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Programmatic access to the Weblens server",
        "title": "Weblens API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/",
    "paths": {
...
        "/tower": {
            "get": {
                "security": [
                    {
                        "SessionAuth": [
                            "admin"
                        ]
                    },
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "tags": [
                    "Towers"
                ],
                "summary": "Get all remotes",
                "operationId": "GetRemotes",
                "responses": {
                    "200": {
                        "description": "Tower Info",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TowerInfo"
                            }
                        }
                    }
                }
            }
        },
        "/tower/init": {
            "post": {
                "security": [],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Towers"
                ],
                "summary": "Initialize the target server",
                "operationId": "InitializeTower",
                "parameters": [
                    {
                        "description": "Server initialization body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.InitServerParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New server info",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TowerInfo"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "security": [
                    {
                        "SessionAuth": [
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get all users, including (possibly) sensitive information like password hashes",
                "operationId": "GetUsers",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserInfoArchive"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "CreateUser",
                "parameters": [
                    {
                        "description": "New user params",
                        "name": "newUserParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewUserParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "NewUserParams": {
            "type": "object",
            "required": [
                "fullName",
                "password",
                "username"
            ],
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "autoActivate": {
                    "type": "boolean"
                },
                "fullName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "TowerInfo": {
            "type": "object",
            "properties": {
                "backupSize": {
                    "type": "integer"
                },
                "coreAddress": {
                    "description": "Address of the remote server, only if the instance is a core.\nNot set for any remotes/backups on core server, as it IS the core",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastBackup": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "online": {
                    "type": "boolean"
                },
                "reportedRole": {
                    "description": "Role the server is currently reporting. This is used to determine if the server is online (and functional) or not",
                    "type": "string"
                },
                "role": {
                    "description": "Core or Backup",
                    "type": "string"
                },
                "started": {
                    "type": "boolean"
                },
                "userCount": {
                    "type": "integer"
                }
            }
        },
...
        "UserInfoArchive": {
            "type": "object",
            "properties": {
                "activated": {
                    "type": "boolean"
                },
                "fullName": {
                    "type": "string"
                },
                "homeId": {
                    "type": "string"
                },
                "homeSize": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "permissionLevel": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "trashId": {
                    "type": "string"
                },
                "trashSize": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "structs.InitServerParams": {
            "type": "object",
            "properties": {
                "coreAddress": {
                    "type": "string"
                },
                "coreKey": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "localId": {
                    "description": "For restoring a server, remoind the core of its serverId and api key the remote last used",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "remoteId": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "usingKeyInfo": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "scopes": {
                "admin": "Grants read and write access to privileged data"
            }
        },
        "SessionAuth": {
            "type": "apiKey",
            "name": "weblens-session-token",
            "in": "cookie"
        }
    }
}